refactor: split types into file
